app/pyproject.toml
[project]
name = "ImIn"
version = "0.0.0"
dependencies = [
    "customtkinter==5.2.2",
    "darkdetect==0.8.0",
    "ImageIO==2.37.2",
    "jaraco.classes==3.4.0",
    "jaraco.context==6.0.2",
    "jaraco.functools==4.4.0",
    "keyring==25.7.0",
    "more-itertools==10.8.0",
    "Nuitka==2.8.9",
    "numpy==2.4.0",
    "ordered-set==4.1.0",
    "packaging==25.0",
    "pillow==12.0.0",
    "pyscard==2.3.1",
    "slack-bolt==1.27.0",
    "slack-sdk==3.39.0",
    "zstandard==0.25.0"
]